requirements.txt
# Bio-Medical AI Competition - Simple Starter Kit

# Core dependencies
transformers==4.54.1
torch>=1.12.0
tqdm>=4.64.0
pandas>=1.3.0

# Utility packages
numpy>=1.21.0
scikit-learn>=1.0.0  # For additional metrics (optional)

# Development dependencies (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# SFT of LLMs
trl==0.19.1
peft==0.16.0
accelerate==1.9.0
bitsandbytes==0.46.1

# Quantized LLM 
llama-cpp-python

